add verification on profile name

+++ back/src/services/profile.service.ts
@@ -51,6 +51,7 @@ export class ProfileService {
     }
     };
   async delete(name : string ){
+    if(name === "Invité"){return {"status":401,"message":"Can't delete guest profile"};}
     let res = await this.profileRepository.deleteOne({pseudo:name});
     let obj={};
     console.log(res.result);
